\name{RcppNumericalSolver-package}
\alias{RcppNumericalSolver-package}
\alias{RcppNumericalSolver}
\docType{package}
\title{\packageTitle{RcppNumericalSolver}}
\description{\packageDescription{RcppNumericalSolver}}
\details{The DESCRIPTION file: \packageDESCRIPTION{RcppNumericalSolver}}
\section{Package Content}{\packageIndices{RcppNumericalSolver}}
\author{\packageAuthor{RcppNumericalSolver}}
\section{Maintainer}{\packageMaintainer{RcppNumericalSolver}}
\references{https://github.com/PatWie/CppNumericalSolvers}
\keyword{ package }
